Chatbot_Earthrenewal/requirements.txt
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
setuptools-rust
git+https://github.com/openai/whisper.git
gtts
googletrans==4.0.0-rc1
pyaudio
playsound
